+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.3] - 2018-12-04
 ### Added
 - `SQLiteDatabase#delete()` method
+- `SQLiteDatabase#update()` method
 
 ## [0.2.2] - 2018-12-04
 ### Added